Crate compositing

Re-exports

pub use crate::compositor_thread::CompositorProxy;

Modules

compositor

compositor_thread
Communication with the compositor thread.

gl

touch

windowing
Abstract windowing methods. The concrete implementations of these can be found in platform/.

Structs

CompositionPipeline
The subset of the pipeline that is needed for layer composition.

IOCompositor
NB: Never block on the constellation, because sometimes the constellation blocks on us.

SendableFrameTree

Enums

CompositingReason
Why we performed a composite. This is used for debugging.

ConstellationMsg
Messages to the constellation.

ShutdownState