Struct compositing::compositor::IOCompositor

pub struct IOCompositor<Window: WindowMethods> {
    window: Rc<Window>,
    port: Box<CompositorReceiver>,
    root_pipeline: Option<CompositionPipeline>,
    pipeline_details: HashMap<PipelineId, PipelineDetails>,
    scale: ScaleFactor<f32, LayerPixel, DevicePixel>,
    window_size: TypedSize2D<u32, DevicePixel>,
    viewport: Option<(TypedPoint2D<u32, DevicePixel>, TypedSize2D<u32, DevicePixel>)>,
    viewport_zoom: ScaleFactor<f32, PagePx, ViewportPx>,
    min_viewport_zoom: Option<ScaleFactor<f32, PagePx, ViewportPx>>,
    max_viewport_zoom: Option<ScaleFactor<f32, PagePx, ViewportPx>>,
    page_zoom: ScaleFactor<f32, ViewportPx, ScreenPx>,
    scale_factor: ScaleFactor<f32, ScreenPx, DevicePixel>,
    channel_to_self: Box<CompositorProxy + Send>,
    delayed_composition_timer: DelayedCompositionTimerProxy,
    composite_target: CompositeTarget,
    composition_request: CompositionRequest,
    shutdown_state: ShutdownState,
    last_composite_time: u64,
    zoom_action: bool,
    zoom_time: f64,
    got_load_complete_message: bool,
    frame_tree_id: FrameTreeId,
    constellation_chan: Sender<ConstellationMsg>,
    time_profiler_chan: ProfilerChan,
    touch_handler: TouchHandler,
    pending_scroll_zoom_events: Vec<ScrollZoomEvent>,
    waiting_for_results_of_scroll: bool,
    ready_to_save_state: ReadyState,
    scroll_in_progress: bool,
    webrender: Renderer,
    webrender_api: RenderApi,
}

NB: Never block on the constellation, because sometimes the constellation blocks on us.

Fields
window: Rc<Window>

The application window.

port: Box<CompositorReceiver>

The port on which we receive messages.

root_pipeline: Option<CompositionPipeline>

The root pipeline.

pipeline_details: HashMap<PipelineId, PipelineDetails>

Tracks details about each active pipeline that the compositor knows about.

scale: ScaleFactor<f32, LayerPixel, DevicePixel>

The scene scale, to allow for zooming and high-resolution painting.

window_size: TypedSize2D<u32, DevicePixel>

The application window size.

viewport: Option<(TypedPoint2D<u32, DevicePixel>, TypedSize2D<u32, DevicePixel>)>

The overridden viewport.

viewport_zoom: ScaleFactor<f32, PagePx, ViewportPx>

"Mobile-style" zoom that does not reflow the page.

min_viewport_zoom: Option<ScaleFactor<f32, PagePx, ViewportPx>>

Viewport zoom constraints provided by @viewport.

max_viewport_zoom: Option<ScaleFactor<f32, PagePx, ViewportPx>> page_zoom: ScaleFactor<f32, ViewportPx, ScreenPx>

"Desktop-style" zoom that resizes the viewport to fit the window. See ViewportPx docs in util/geom.rs for details.

scale_factor: ScaleFactor<f32, ScreenPx, DevicePixel>

The device pixel ratio for this window.

channel_to_self: Box<CompositorProxy + Send> delayed_composition_timer: DelayedCompositionTimerProxy

A handle to the delayed composition timer.

composite_target: CompositeTarget

The type of composition to perform

composition_request: CompositionRequest

Tracks whether we should composite this frame.

shutdown_state: ShutdownState

Tracks whether we are in the process of shutting down, or have shut down and should close the compositor.

last_composite_time: u64

Tracks the last composite time.

zoom_action: bool

Tracks whether the zoom action has happened recently.

zoom_time: f64

The time of the last zoom action has started.

got_load_complete_message: bool

Whether the page being rendered has loaded completely. Differs from ReadyState because we can finish loading (ready) many times for a single page.

frame_tree_id: FrameTreeId

The current frame tree ID (used to reject old paint buffers)

constellation_chan: Sender<ConstellationMsg>

The channel on which messages can be sent to the constellation.

time_profiler_chan: ProfilerChan

The channel on which messages can be sent to the time profiler.

touch_handler: TouchHandler

Touch input state machine

pending_scroll_zoom_events: Vec<ScrollZoomEvent>

Pending scroll/zoom events.

waiting_for_results_of_scroll: bool

Whether we're waiting on a recomposite after dispatching a scroll.

ready_to_save_state: ReadyState

Used by the logic that determines when it is safe to output an image for the reftest framework.

scroll_in_progress: bool

Whether a scroll is in progress; i.e. whether the user's fingers are down.

webrender: Renderer

The webrender renderer.

webrender_api: RenderApi

The webrender interface, if enabled.

Methods

impl<Window: WindowMethods> IOCompositor<Window>

fn new(window: Rc<Window>,
state: InitialCompositorState)
-> IOCompositor<Window>

fn create(window: Rc<Window>,
state: InitialCompositorState)
-> IOCompositor<Window>

fn start_shutting_down(&mut self)

fn finish_shutting_down(&mut self)

fn handle_browser_message(&mut self, msg: Msg) -> bool

fn change_running_animations_state(&mut self,
pipeline_id: PipelineId,
animation_state: AnimationState)

Sets or unsets the animations-running flag for the given pipeline, and schedules a recomposite if necessary.

fn pipeline_details(&mut self, pipeline_id: PipelineId) -> &mut PipelineDetails

fn pipeline(&self, pipeline_id: PipelineId) -> Option<&CompositionPipeline>

fn change_page_title(&mut self, pipeline_id: PipelineId, title: Option<String>)

fn change_page_url(&mut self, _: PipelineId, url: ServoUrl)

fn set_frame_tree(&mut self,
frame_tree: &SendableFrameTree,
response_chan: IpcSender<()>)

fn create_pipeline_details_for_frame_tree(&mut self,
frame_tree: &SendableFrameTree)

fn remove_pipeline_root_layer(&mut self, pipeline_id: PipelineId)

fn send_window_size(&self, size_type: WindowSizeType)

fn schedule_delayed_composite_if_necessary(&mut self)

fn scroll_fragment_to_point(&mut self,
_pipeline_id: PipelineId,
_point: Point2D<f32>)

fn handle_window_message(&mut self, event: WindowEvent)

fn on_resize_window_event(&mut self, new_size: TypedSize2D<u32, DevicePixel>)

fn on_load_url_window_event(&mut self, url_string: String)

fn on_mouse_window_event_class(&mut self, mouse_window_event: MouseWindowEvent)

fn dispatch_mouse_window_event_class(&mut self,
mouse_window_event: MouseWindowEvent)

fn on_mouse_window_move_event_class(&mut self,
cursor: TypedPoint2D<f32, DevicePixel>)

fn dispatch_mouse_window_move_event_class(&mut self,
cursor: TypedPoint2D<f32, DevicePixel>)

fn send_event_to_root_pipeline(&self, event: CompositorEvent)

fn on_touch_down(&mut self,
identifier: TouchId,
point: TypedPoint2D<f32, DevicePixel>)

fn on_touch_move(&mut self,
identifier: TouchId,
point: TypedPoint2D<f32, DevicePixel>)

fn on_touch_up(&mut self,
identifier: TouchId,
point: TypedPoint2D<f32, DevicePixel>)

fn on_touch_cancel(&mut self,
identifier: TouchId,
point: TypedPoint2D<f32, DevicePixel>)

fn on_touchpad_pressure_event(&self,
point: TypedPoint2D<f32, DevicePixel>,
pressure: f32,
phase: TouchpadPressurePhase)

fn simulate_mouse_click(&mut self, p: TypedPoint2D<f32, DevicePixel>)

http://w3c.github.io/touch-events/#mouse-events

fn on_scroll_window_event(&mut self,
delta: TypedPoint2D<f32, DevicePixel>,
cursor: TypedPoint2D<i32, DevicePixel>)

fn on_scroll_start_window_event(&mut self,
delta: TypedPoint2D<f32, DevicePixel>,
cursor: TypedPoint2D<i32, DevicePixel>)

fn on_scroll_end_window_event(&mut self,
delta: TypedPoint2D<f32, DevicePixel>,
cursor: TypedPoint2D<i32, DevicePixel>)

fn process_pending_scroll_events(&mut self)

fn process_animations(&mut self)

If there are any animations running, dispatches appropriate messages to the constellation.

fn tick_animations_for_pipeline(&mut self, pipeline_id: PipelineId)

fn constrain_viewport(&mut self,
pipeline_id: PipelineId,
constraints: ViewportConstraints)

fn device_pixels_per_screen_px(&self) -> ScaleFactor<f32, ScreenPx, DevicePixel>

fn device_pixels_per_page_px(&self) -> ScaleFactor<f32, PagePx, DevicePixel>

fn update_zoom_transform(&mut self)

fn on_zoom_reset_window_event(&mut self)

fn on_zoom_window_event(&mut self, magnification: f32)

fn on_pinch_zoom_window_event(&mut self, magnification: f32)

Simulate a pinch zoom

fn on_navigation_window_event(&self, direction: WindowNavigateMsg)

fn on_key_event(&mut self,
ch: Option<char>,
key: Key,
state: KeyState,
modifiers: KeyModifiers)

fn send_viewport_rects(&self)

fn animations_active(&self) -> bool

fn is_ready_to_paint_image_output(&mut self) -> Result<(), NotReadyToPaint>

Query the constellation to see if the current compositor output matches the current frame tree output, and if the associated script threads are idle.

fn composite(&mut self)

fn composite_specific_target(&mut self,
target: CompositeTarget)
-> Result<Option<Image>, UnableToComposite>

Composite either to the screen or to a png image or both. Returns Ok if composition was performed or Err if it was not possible to composite for some reason. If CompositeTarget is Window or Png no image data is returned; in the latter case the image is written directly to a file. If CompositeTarget is WindowAndPng Ok(Some(png::Image)) is returned.

fn draw_img(&self,
render_target_info: RenderTargetInfo,
width: usize,
height: usize)
-> RgbImage

fn composite_if_necessary(&mut self, reason: CompositingReason)

fn initialize_compositing(&mut self)

fn get_root_pipeline_id(&self) -> Option<PipelineId>

fn start_scrolling_bounce_if_necessary(&mut self)

fn handle_events(&mut self, messages: Vec<WindowEvent>) -> bool

fn repaint_synchronously(&mut self)

Repaints and recomposites synchronously. You must be careful when calling this, as if a paint is not scheduled the compositor will hang forever.

This is used when resizing the window.

fn pinch_zoom_level(&self) -> f32

fn title_for_main_frame(&self)