Trait compositing::compositor_layer::CompositorLayer

pub trait CompositorLayer {
    fn update_layer_except_bounds(&self, layer_properties: LayerProperties);
    fn update_layer(&self, layer_properties: LayerProperties);
    fn add_buffers<Window>(&self, compositor: &mut IOCompositor<Window>, new_buffers: Box<LayerBufferSet>, epoch: Epoch) where Window: WindowMethods;
    fn clear<Window>(&self, compositor: &mut IOCompositor<Window>) where Window: WindowMethods;
    fn clear_all_tiles<Window>(&self, compositor: &mut IOCompositor<Window>) where Window: WindowMethods;
    fn remove_root_layer_with_pipeline_id<Window>(&self, compositor: &mut IOCompositor<Window>, pipeline_id: PipelineId) where Window: WindowMethods;
    fn forget_all_tiles(&self);
    fn handle_scroll_event(&self, delta: TypedPoint2D<LayerPixel, f32>, cursor: TypedPoint2D<LayerPixel, f32>) -> ScrollEventResult;
    fn send_mouse_event<Window>(&self, compositor: &IOCompositor<Window>, event: MouseWindowEvent, cursor: TypedPoint2D<LayerPixel, f32>) where Window: WindowMethods;
    fn send_mouse_move_event<Window>(&self, compositor: &IOCompositor<Window>, cursor: TypedPoint2D<LayerPixel, f32>) where Window: WindowMethods;
    fn send_event<Window>(&self, compositor: &IOCompositor<Window>, event: CompositorEvent) where Window: WindowMethods;
    fn send_touchpad_pressure_event<Window>(&self, compositor: &IOCompositor<Window>, cursor: TypedPoint2D<LayerPixel, f32>, pressure: f32, phase: TouchpadPressurePhase) where Window: WindowMethods;
    fn clamp_scroll_offset_and_scroll_layer(&self, new_offset: TypedPoint2D<LayerPixel, f32>) -> ScrollEventResult;
    fn scroll_layer_and_all_child_layers(&self, new_offset: TypedPoint2D<LayerPixel, f32>) -> bool;
    fn wants_scroll_events(&self) -> WantsScrollEventsFlag;
    fn pipeline_id(&self) -> PipelineId;
}

Required Methods

fn update_layer_except_bounds(&self, layer_properties: LayerProperties)

fn update_layer(&self, layer_properties: LayerProperties)

fn add_buffers<Window>(&self, compositor: &mut IOCompositor<Window>, new_buffers: Box<LayerBufferSet>, epoch: Epoch) where Window: WindowMethods

fn clear<Window>(&self, compositor: &mut IOCompositor<Window>) where Window: WindowMethods

Destroys all layer tiles, sending the buffers back to the painter to be destroyed or reused.

fn clear_all_tiles<Window>(&self, compositor: &mut IOCompositor<Window>) where Window: WindowMethods

Destroys tiles for this layer and all descendent layers, sending the buffers back to the painter to be destroyed or reused.

fn remove_root_layer_with_pipeline_id<Window>(&self, compositor: &mut IOCompositor<Window>, pipeline_id: PipelineId) where Window: WindowMethods

Removes the root layer (and any children) for a given pipeline from the compositor. Buffers that the compositor is holding are returned to the owning paint thread.

fn forget_all_tiles(&self)

Destroys all tiles of all layers, including children, without sending them back to the painter. You must call this only when the paint thread is destined to be going down; otherwise, you will leak tiles.

This is used during shutdown, when we know the paint thread is going away.

fn handle_scroll_event(&self, delta: TypedPoint2D<LayerPixel, f32>, cursor: TypedPoint2D<LayerPixel, f32>) -> ScrollEventResult

Move the layer's descendants that don't want scroll events and scroll by a relative specified amount in page coordinates. This also takes in a cursor position to see if the mouse is over child layers first. If a layer successfully scrolled returns either ScrollPositionUnchanged or ScrollPositionChanged. If no layer was targeted by the event returns ScrollEventUnhandled.

fn send_mouse_event<Window>(&self, compositor: &IOCompositor<Window>, event: MouseWindowEvent, cursor: TypedPoint2D<LayerPixel, f32>) where Window: WindowMethods

fn send_mouse_move_event<Window>(&self, compositor: &IOCompositor<Window>, cursor: TypedPoint2D<LayerPixel, f32>) where Window: WindowMethods

fn send_event<Window>(&self, compositor: &IOCompositor<Window>, event: CompositorEvent) where Window: WindowMethods

fn send_touchpad_pressure_event<Window>(&self, compositor: &IOCompositor<Window>, cursor: TypedPoint2D<LayerPixel, f32>, pressure: f32, phase: TouchpadPressurePhase) where Window: WindowMethods

fn clamp_scroll_offset_and_scroll_layer(&self, new_offset: TypedPoint2D<LayerPixel, f32>) -> ScrollEventResult

fn scroll_layer_and_all_child_layers(&self, new_offset: TypedPoint2D<LayerPixel, f32>) -> bool

fn wants_scroll_events(&self) -> WantsScrollEventsFlag

Return a flag describing how this layer deals with scroll events.

fn pipeline_id(&self) -> PipelineId

Return the pipeline id associated with this layer.