Module compositing::compositor_task

Communication with the compositor task.

Reexports

pub use windowing;

pub use constellation::SendableFrameTree;

Structs

CompositorTask

LayerProperties
Information about each layer that the compositor keeps.

Enums

Msg
Messages from the painting task and the constellation task to the compositor task.

Traits

CompositorEventListener

CompositorProxy
Sends messages to the compositor. This is a trait supplied by the port because the method used to communicate with the compositor may have to kick OS event loops awake, communicate cross- process, and so forth.

CompositorReceiver
The port that the compositor receives messages on. As above, this is a trait supplied by the Servo port.