Enum compositing::compositor_task::Msg

pub enum Msg {
    Exit(Sender<()>),
    ShutdownComplete,
    GetGraphicsMetadata(Sender<Option<NativeGraphicsMetadata>>),
    CreateOrUpdateBaseLayer(LayerProperties),
    CreateOrUpdateDescendantLayer(LayerProperties),
    SetLayerRect(PipelineId, LayerId, Rect<f32>),
    ScrollFragmentPoint(PipelineId, LayerId, Point2D<f32>),
    AssignPaintedBuffers(PipelineId, Epoch, Vec<(LayerId, Box<LayerBufferSet>)>),
    ChangeReadyState(PipelineId, ReadyState),
    ChangePaintState(PipelineId, PaintState),
    ChangePageTitle(PipelineId, Option<String>),
    ChangePageUrl(PipelineId, Url),
    ChangeRunningAnimationsState(PipelineId, bool),
    PaintMsgDiscarded,
    SetFrameTree(SendableFrameTree, Sender<()>, ConstellationChan),
    LoadComplete,
    ScrollTimeout(u64),
    KeyEvent(Key, KeyState, KeyModifiers),
    SetCursor(Cursor),
    PaintTaskExited(PipelineId),
}

Messages from the painting task and the constellation task to the compositor task.

Variants

Exit

Requests that the compositor shut down.

ShutdownComplete

Informs the compositor that the constellation has completed shutdown. Required because the constellation can have pending calls to make (e.g. SetFrameTree) at the time that we send it an ExitMsg.

GetGraphicsMetadata

Requests the compositor's graphics metadata. Graphics metadata is what the painter needs to create surfaces that the compositor can see. On Linux this is the X display; on Mac this is the pixel format.

The headless compositor returns None.

CreateOrUpdateBaseLayer

Tells the compositor to create the root layer for a pipeline if necessary (i.e. if no layer with that ID exists).

CreateOrUpdateDescendantLayer

Tells the compositor to create a descendant layer for a pipeline if necessary (i.e. if no layer with that ID exists).

SetLayerRect

Alerts the compositor that the specified layer's rect has changed.

ScrollFragmentPoint

Scroll a page in a window

AssignPaintedBuffers

Requests that the compositor assign the painted buffers to the given layers.

ChangeReadyState

Alerts the compositor to the current status of page loading.

ChangePaintState

Alerts the compositor to the current status of painting.

ChangePageTitle

Alerts the compositor that the current page has changed its title.

ChangePageUrl

Alerts the compositor that the current page has changed its URL.

ChangeRunningAnimationsState

Alerts the compositor that the given pipeline has changed whether it is running animations.

PaintMsgDiscarded

Alerts the compositor that a PaintMsg has been discarded.

SetFrameTree

Replaces the current frame tree, typically called during main frame navigation.

LoadComplete

The load of a page has completed.

ScrollTimeout

Indicates that the scrolling timeout with the given starting timestamp has happened and a composite should happen. (See the scrolling module.)

KeyEvent

Sends an unconsumed key event back to the compositor.

SetCursor

Changes the cursor.

PaintTaskExited

Informs the compositor that the paint task for the given pipeline has exited.