Module compositing::compositor_thread

Communication with the compositor thread.

Structs

CompositorThread

InitialCompositorState
Data used to construct a compositor.

Enums

Msg
Messages from the painting thread and the constellation thread to the compositor thread.

Traits

CompositorEventListener

CompositorProxy
Sends messages to the compositor. This is a trait supplied by the port because the method used to communicate with the compositor may have to kick OS event loops awake, communicate cross- process, and so forth.

CompositorReceiver
The port that the compositor receives messages on. As above, this is a trait supplied by the Servo port.

RenderListener