Enum compositing::compositor_thread::Msg

pub enum Msg {
    Exit,
    ShutdownComplete,
    ScrollFragmentPoint(PipelineId, ScrollRootId, Point2D<f32>, bool),
    ChangePageTitle(PipelineId, Option<String>),
    ChangePageUrl(PipelineId, ServoUrl),
    ChangeRunningAnimationsState(PipelineId, AnimationState),
    SetFrameTree(SendableFrameTree, IpcSender<()>),
    LoadStart(bool, bool),
    LoadComplete(bool, bool, bool),
    DelayedCompositionTimeout(u64),
    Recomposite(CompositingReason),
    KeyEvent(Option<char>, Key, KeyState, KeyModifiers),
    TouchEventProcessed(EventResult),
    SetCursor(Cursor),
    CreatePng(IpcSender<Option<Image>>),
    ViewportConstrained(PipelineId, ViewportConstraints),
    IsReadyToSaveImageReply(bool),
    NewFavicon(ServoUrl),
    HeadParsed,
    Status(Option<String>),
    GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>),
    MoveTo(Point2D<i32>),
    ResizeTo(Size2D<u32>),
    PipelineVisibilityChanged(PipelineId, bool),
    NewScrollFrameReady(bool),
    PipelineExited(PipelineId, IpcSender<()>),
    Dispatch(Box<Fn() + Send>),
    SetFullscreenState(bool),
}

Messages from the painting thread and the constellation thread to the compositor thread.

Variants
Exit

Requests that the compositor shut down.

ShutdownComplete

Informs the compositor that the constellation has completed shutdown. Required because the constellation can have pending calls to make (e.g. SetFrameTree) at the time that we send it an ExitMsg.

ScrollFragmentPoint(PipelineId, ScrollRootId, Point2D<f32>, bool)

Scroll a page in a window

ChangePageTitle(PipelineId, Option<String>)

Alerts the compositor that the current page has changed its title.

ChangePageUrl(PipelineId, ServoUrl)

Alerts the compositor that the current page has changed its URL.

ChangeRunningAnimationsState(PipelineId, AnimationState)

Alerts the compositor that the given pipeline has changed whether it is running animations.

SetFrameTree(SendableFrameTree, IpcSender<()>)

Replaces the current frame tree, typically called during main frame navigation.

LoadStart(bool, bool)

The load of a page has begun: (can go back, can go forward).

LoadComplete(bool, bool, bool)

The load of a page has completed: (can go back, can go forward, is root frame).

DelayedCompositionTimeout(u64)

We hit the delayed composition timeout. (See delayed_composition.rs.)

Recomposite(CompositingReason)

Composite.

KeyEvent(Option<char>, Key, KeyState, KeyModifiers)

Sends an unconsumed key event back to the compositor.

TouchEventProcessed(EventResult)

Script has handled a touch event, and either prevented or allowed default actions.

SetCursor(Cursor)

Changes the cursor.

CreatePng(IpcSender<Option<Image>>)

Composite to a PNG file and return the Image over a passed channel.

ViewportConstrained(PipelineId, ViewportConstraints)

Alerts the compositor that the viewport has been constrained in some manner

IsReadyToSaveImageReply(bool)

A reply to the compositor asking if the output image is stable.

NewFavicon(ServoUrl)

A favicon was detected

HeadParsed

Status(Option<String>)

A status message to be displayed by the browser chrome.

GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>)

Get Window Informations size and position

MoveTo(Point2D<i32>)

Move the window to a point

ResizeTo(Size2D<u32>)

Resize the window to size

PipelineVisibilityChanged(PipelineId, bool)

Pipeline visibility changed

NewScrollFrameReady(bool)

WebRender has successfully processed a scroll. The boolean specifies whether a composite is needed.

PipelineExited(PipelineId, IpcSender<()>)

A pipeline was shut down.

Dispatch(Box<Fn() + Send>)

Runs a closure in the compositor thread. It's used to dispatch functions from webrender to the main thread's event loop. Required to allow WGL GLContext sharing in Windows.

SetFullscreenState(bool)

Enter or exit fullscreen