Struct compositing::constellation::Constellation

pub struct Constellation<LTF, STF> {
    pub chan: ConstellationChan,
    pub request_port: Receiver<ConstellationMsg>,
    pub compositor_proxy: Box<CompositorProxy>,
    pub resource_task: ResourceTask,
    pub image_cache_task: ImageCacheTask,
    pub time_profiler_chan: ProfilerChan,
    pub mem_profiler_chan: ProfilerChan,
    pub window_size: WindowSizeData,
    // some fields omitted
}

Maintains the pipelines and navigation context and grants permission to composite.

Fields

chan

A channel through which messages can be sent to this object.

request_port

Receives messages.

compositor_proxy

A channel (the implementation of which is port-specific) through which messages can be sent to the compositor.

resource_task

A channel through which messages can be sent to the resource task.

image_cache_task

A channel through which messages can be sent to the image cache task.

time_profiler_chan

A channel through which messages can be sent to the time profiler.

mem_profiler_chan

A channel through which messages can be sent to the memory profiler.

window_size

Methods

impl<LTF: LayoutTaskFactory, STF: ScriptTaskFactory> Constellation<LTF, STF>

fn start(compositor_proxy: Box<CompositorProxy + Send>, resource_task: ResourceTask, image_cache_task: ImageCacheTask, font_cache_task: FontCacheTask, time_profiler_chan: ProfilerChan, mem_profiler_chan: ProfilerChan, devtools_chan: Option<DevtoolsControlChan>, storage_task: StorageTask) -> ConstellationChan