Enum compositing::CompositorMsg

pub enum CompositorMsg {
    Exit,
    FrameSize(PipelineId, Size2D<f32>),
    GetFrame(PipelineId, IpcSender<Option<FrameId>>),
    GetPipeline(Option<FrameId>, IpcSender<Option<(PipelineId, bool)>>),
    GetPipelineTitle(PipelineId),
    InitLoadUrl(Url),
    IsReadyToSaveImage(HashMap<PipelineId, Epoch>),
    KeyEvent(Key, KeyState, KeyModifiers),
    LoadUrl(PipelineId, LoadData),
    Navigate(Option<(PipelineId, SubpageId)>, NavigationDirection),
    WindowSize(WindowSizeData, WindowSizeType),
    TickAnimation(PipelineId, AnimationTickType),
    WebDriverCommand(WebDriverCommandMsg),
}

Messages from the compositor to the constellation.

Variants

Exit

FrameSize

GetFrame

Request that the constellation send the FrameId corresponding to the document with the provided pipeline id

GetPipeline

Request that the constellation send the current pipeline id for the provided frame id, or for the root frame if this is None, over a provided channel. Also returns a boolean saying whether the document has finished loading or not.

GetPipelineTitle

Requests that the constellation inform the compositor of the title of the pipeline immediately.

InitLoadUrl

IsReadyToSaveImage

Query the constellation to see if the current compositor output is stable

KeyEvent

LoadUrl

Navigate

WindowSize

TickAnimation

Requests that the constellation instruct layout to begin a new tick of the animation.

WebDriverCommand

Dispatch a webdriver command