Enum compositing::ConstellationMsg

pub enum ConstellationMsg {
    Exit,
    GetBrowsingContext(PipelineId, IpcSender<Option<BrowsingContextId>>),
    GetPipeline(BrowsingContextId, IpcSender<Option<PipelineId>>),
    GetFocusTopLevelBrowsingContext(IpcSender<Option<TopLevelBrowsingContextId>>),
    IsReadyToSaveImage(HashMap<PipelineId, Epoch>),
    Keyboard(KeyboardEvent),
    AllowNavigationResponse(PipelineId, bool),
    LoadUrl(TopLevelBrowsingContextId, ServoUrl),
    TraverseHistory(TopLevelBrowsingContextId, TraversalDirection),
    WindowSize(Option<TopLevelBrowsingContextId>, WindowSizeData, WindowSizeType),
    TickAnimation(PipelineId, AnimationTickType),
    WebDriverCommand(WebDriverCommandMsg),
    Reload(TopLevelBrowsingContextId),
    LogEntry(Option<TopLevelBrowsingContextId>, Option<String>, LogEntry),
    WebVREvents(Vec<PipelineId>, Vec<WebVREvent>),
    NewBrowser(ServoUrl, TopLevelBrowsingContextId),
    CloseBrowser(TopLevelBrowsingContextId),
    SendError(Option<TopLevelBrowsingContextId>, String),
    SelectBrowser(TopLevelBrowsingContextId),
    ForwardEvent(PipelineId, CompositorEvent),
    SetCursor(Cursor),
    EnableProfiler(Duration, Duration),
    DisableProfiler,
    ExitFullScreen(TopLevelBrowsingContextId),
    MediaSessionAction(MediaSessionActionType),
    ChangeBrowserVisibility(TopLevelBrowsingContextId, bool),
}

Messages to the constellation.

Variants

Exit

Exit the constellation.

GetBrowsingContext(PipelineId, IpcSender<Option<BrowsingContextId>>)

Request that the constellation send the BrowsingContextId corresponding to the document with the provided pipeline id

GetPipeline(BrowsingContextId, IpcSender<Option<PipelineId>>)

Request that the constellation send the current pipeline id for the provided browsing context id, over a provided channel.

GetFocusTopLevelBrowsingContext(IpcSender<Option<TopLevelBrowsingContextId>>)

Request that the constellation send the current focused top-level browsing context id, over a provided channel.

IsReadyToSaveImage(HashMap<PipelineId, Epoch>)

Query the constellation to see if the current compositor output is stable

Keyboard(KeyboardEvent)

Inform the constellation of a key event.

AllowNavigationResponse(PipelineId, bool)

Whether to allow script to navigate.

LoadUrl(TopLevelBrowsingContextId, ServoUrl)

Request to load a page.

TraverseHistory(TopLevelBrowsingContextId, TraversalDirection)

Request to traverse the joint session history of the provided browsing context.

WindowSize(Option<TopLevelBrowsingContextId>, WindowSizeData, WindowSizeType)

Inform the constellation of a window being resized.

TickAnimation(PipelineId, AnimationTickType)

Requests that the constellation instruct layout to begin a new tick of the animation.

WebDriverCommand(WebDriverCommandMsg)

Dispatch a webdriver command

Reload(TopLevelBrowsingContextId)

Reload a top-level browsing context.

LogEntry(Option<TopLevelBrowsingContextId>, Option<String>, LogEntry)

A log entry, with the top-level browsing context id and thread name

WebVREvents(Vec<PipelineId>, Vec<WebVREvent>)

Dispatch WebVR events to the subscribed script threads.

NewBrowser(ServoUrl, TopLevelBrowsingContextId)

Create a new top level browsing context.

CloseBrowser(TopLevelBrowsingContextId)

Close a top level browsing context.

SendError(Option<TopLevelBrowsingContextId>, String)

Panic a top level browsing context.

SelectBrowser(TopLevelBrowsingContextId)

Make browser visible.

ForwardEvent(PipelineId, CompositorEvent)

Forward an event to the script task of the given pipeline.

SetCursor(Cursor)

Requesting a change to the onscreen cursor.

EnableProfiler(Duration, Duration)

Enable the sampling profiler, with a given sampling rate and max total sampling duration.

DisableProfiler

Disable the sampling profiler.

ExitFullScreen(TopLevelBrowsingContextId)

Request to exit from fullscreen mode

MediaSessionAction(MediaSessionActionType)

Media session action.

ChangeBrowserVisibility(TopLevelBrowsingContextId, bool)

Toggle browser visibility.