Struct compositing::pipeline::Pipeline

pub struct Pipeline {
    pub id: PipelineId,
    pub parent_info: Option<(PipelineId, SubpageId)>,
    pub script_chan: ScriptControlChan,
    pub layout_chan: LayoutControlChan,
    pub paint_chan: PaintChan,
    pub layout_shutdown_port: Receiver<()>,
    pub paint_shutdown_port: Receiver<()>,
    pub url: Url,
    pub title: Option<String>,
    pub rect: Option<TypedRect<PagePx, f32>>,
    pub running_animations: bool,
    pub children: Vec<FrameId>,
}

A uniquely-identifiable pipeline of script task, layout task, and paint task.

Fields

id

parent_info

script_chan

layout_chan

A channel to layout, for performing reflows and shutdown.

paint_chan

layout_shutdown_port

paint_shutdown_port

url

URL corresponding to the most recently-loaded page.

title

The title of the most recently-loaded page.

rect

running_animations

Whether this pipeline is currently running animations. Pipelines that are running animations cause composites to be continually scheduled.

children

Methods

impl Pipeline

fn create<LTF, STF>(id: PipelineId, parent_info: Option<(PipelineId, SubpageId)>, constellation_chan: ConstellationChan, compositor_proxy: Box<CompositorProxy + 'static + Send>, devtools_chan: Option<DevtoolsControlChan>, image_cache_task: ImageCacheTask, font_cache_task: FontCacheTask, resource_task: ResourceTask, storage_task: StorageTask, time_profiler_chan: ProfilerChan, mem_profiler_chan: ProfilerChan, window_rect: Option<TypedRect<PagePx, f32>>, script_chan: Option<ScriptControlChan>, load_data: LoadData, device_pixel_ratio: ScaleFactor<ViewportPx, DevicePixel, f32>) -> Pipeline where LTF: LayoutTaskFactory, STF: ScriptTaskFactory

Starts a paint task, layout task, and possibly a script task. Returns the channels wrapped in a struct. If script_pipeline is not None, then subpage_id must also be not None.

fn new(id: PipelineId, parent_info: Option<(PipelineId, SubpageId)>, script_chan: ScriptControlChan, layout_chan: LayoutControlChan, paint_chan: PaintChan, layout_shutdown_port: Receiver<()>, paint_shutdown_port: Receiver<()>, url: Url, rect: Option<TypedRect<PagePx, f32>>) -> Pipeline

fn grant_paint_permission(&self)

fn revoke_paint_permission(&self)

fn exit(&self, exit_type: PipelineExitType)

fn freeze(&self)

fn thaw(&self)

fn force_exit(&self)

fn to_sendable(&self) -> CompositionPipeline

fn add_child(&mut self, frame_id: FrameId)

fn trigger_mozbrowser_event(&self, subpage_id: SubpageId, event: MozBrowserEvent)