Struct compositing::surface_map::SurfaceMap

pub struct SurfaceMap {
    map: HashMap<SurfaceKey, SurfaceValue>,
    mem: usize,
    max_mem: usize,
    counter: usize,
}

This is a struct used to store surfaces when they are not in use. The paint thread can quickly query for a particular size of surface when it needs it.

Fields

map

A HashMap that stores the Buffers.

mem

The current amount of memory stored by the SurfaceMap's surfaces.

max_mem

The maximum allowed memory. Unused surfaces will be deleted when this threshold is exceeded.

counter

A monotonically increasing counter to track how recently tile sizes were used.

Methods

impl SurfaceMap

fn new(max_mem: usize) -> SurfaceMap

fn insert_surfaces<I>(&mut self, display: &NativeDisplay, surfaces: I) where I: IntoIterator<Item=NativeSurface>

fn insert(&mut self, display: &NativeDisplay, new_surface: NativeSurface)

Insert a new buffer into the map.

fn find(&mut self, size: Size2D<i32>) -> Option<NativeSurface>

fn mem(&self) -> usize