compositing

Module touch

Structs

FlingAction 🔒

FlingRefreshDriverObserver 🔒

HitTestResultCache 🔒

A cached PaintHitTestResult to use during a touch sequence. This is kept so that the renderer doesn’t have to constantly keep making hit tests while during panning and flinging actions.

PendingTouchInputEvent 🔒

This data structure is used to store information about touch events that are sent from the Renderer to the Constellation, so that they can finish processing once their DOM events are fired.

TouchHandler

TouchPoint

An action that can be immediately performed in response to a touch move event without waiting for script.

TouchSequenceId 🔒

An ID for a sequence of touch events between a Down and the Up or Cancel event. The ID is the same for all events between Down and Up or Cancel

TouchSequenceInfo

Enums

TouchMoveAllowed

Whether the default move action is allowed or not.

TouchSequenceState 🔒

The states of the touch input state machine.

Constants

FLING_MAX_SCREEN_PX 🔒

Maximum velocity when flinging.

FLING_MIN_SCREEN_PX 🔒

Minimum velocity required for transitioning to fling when panning ends.

FLING_SCALING_FACTOR 🔒

Factor by which the flinging velocity changes on each tick.

TOUCH_PAN_MIN_SCREEN_PX 🔒

Minimum number of DeviceIndependentPixel to begin touch scrolling.