Enum compositing::touch::TouchAction

pub enum TouchAction {
    Click,
    Scroll(TypedPoint2D<DevicePixel, f32>),
    Zoom(f32, TypedPoint2D<DevicePixel, f32>),
    DispatchEvent,
    NoAction,
}

The action to take in response to a touch event

Variants
Click

Simulate a mouse click.

Scroll(TypedPoint2D<DevicePixel, f32>)

Scroll by the provided offset.

Zoom(f32, TypedPoint2D<DevicePixel, f32>)

Zoom by a magnification factor and scroll by the provided offset.

DispatchEvent

Send a JavaScript event to content.

NoAction

Don't do anything.