compositing::touch

Struct TouchHandler

pub struct TouchHandler {
    webview_id: WebViewId,
    pub current_sequence_id: TouchSequenceId,
    touch_sequence_map: FxHashMap<TouchSequenceId, TouchSequenceInfo>,
    pub(crate) pending_touch_input_events: RefCell<FxHashMap<InputEventId, PendingTouchInputEvent>>,
    observing_frames_for_fling: Cell<bool>,
}

Fields
webview_id: WebViewId

The WebViewId of the WebView this TouchHandler is associated with.

current_sequence_id: TouchSequenceIdtouch_sequence_map: FxHashMap<TouchSequenceId, TouchSequenceInfo>pending_touch_input_events: RefCell<FxHashMap<InputEventId, PendingTouchInputEvent>>

A set of InputEventIds for touch events that have been sent to the Constellation and have not been handled yet.

observing_frames_for_fling: Cell<bool>

Whether or not the FlingRefreshDriverObserver is currently observing frames for fling.

Implementations

impl TouchHandler

pub fn new(webview_id: WebViewId) -> Self

pub(crate) fn set_handling_touch_move( &mut self, sequence_id: TouchSequenceId, flag: bool, )

pub(crate) fn is_handling_touch_move( &self, sequence_id: TouchSequenceId, ) -> bool

pub(crate) fn prevent_click(&mut self, sequence_id: TouchSequenceId)

pub(crate) fn prevent_move(&mut self, sequence_id: TouchSequenceId)

pub(crate) fn move_allowed(&mut self, sequence_id: TouchSequenceId) -> bool

Returns true if default move actions are allowed, false if prevented or the result is still pending.,

pub(crate) fn take_pending_touch_move_actions( &mut self, sequence_id: TouchSequenceId, ) -> Vec<ScrollZoomEvent>

pub(crate) fn remove_pending_touch_move_actions( &mut self, sequence_id: TouchSequenceId, )

pub(crate) fn try_remove_touch_sequence(&mut self, sequence_id: TouchSequenceId)

pub(crate) fn remove_touch_sequence(&mut self, sequence_id: TouchSequenceId)

pub fn get_current_touch_sequence_mut(&mut self) -> &mut TouchSequenceInfo

fn try_get_current_touch_sequence(&self) -> Option<&TouchSequenceInfo>

fn try_get_current_touch_sequence_mut( &mut self, ) -> Option<&mut TouchSequenceInfo>

pub(crate) fn get_touch_sequence( &self, sequence_id: TouchSequenceId, ) -> &TouchSequenceInfo

pub(crate) fn get_touch_sequence_mut( &mut self, sequence_id: TouchSequenceId, ) -> Option<&mut TouchSequenceInfo>

pub fn on_touch_down(&mut self, id: TouchId, point: Point2D<f32, DevicePixel>)

pub fn notify_new_frame_start(&mut self) -> Option<FlingAction>

pub(crate) fn stop_fling_if_needed(&mut self)

pub fn on_touch_move( &mut self, id: TouchId, point: Point2D<f32, DevicePixel>, ) -> Option<ScrollZoomEvent>

pub fn on_touch_up(&mut self, id: TouchId, point: Point2D<f32, DevicePixel>)

pub fn on_touch_cancel( &mut self, id: TouchId, _point: Point2D<f32, DevicePixel>, )

pub(crate) fn get_hit_test_result_cache_value( &self, ) -> Option<PaintHitTestResult>

pub(crate) fn set_hit_test_result_cache_value( &mut self, value: PaintHitTestResult, device_pixels_per_page: Scale<f32, CSSPixel, DevicePixel>, )

pub(crate) fn add_pending_touch_input_event( &self, id: InputEventId, event_type: TouchEventType, )

pub(crate) fn take_pending_touch_input_event( &self, id: InputEventId, ) -> Option<PendingTouchInputEvent>

pub(crate) fn add_touch_move_refresh_observer_if_necessary( &self, refresh_driver: Rc<BaseRefreshDriver>, repaint_reason: &Cell<RepaintReason>, )