Enum compositing::windowing::WindowEvent

pub enum WindowEvent {
    Idle,
    Refresh,
    InitializeCompositing,
    Resize(TypedSize2D<DevicePixel, u32>),
    LoadUrl(String),
    MouseWindowEventClass(MouseWindowEvent),
    MouseWindowMoveEventClass(TypedPoint2D<DevicePixel, f32>),
    Scroll(TypedPoint2D<DevicePixel, f32>, TypedPoint2D<DevicePixel, i32>),
    Zoom(f32),
    PinchZoom(f32),
    Navigation(WindowNavigateMsg),
    Quit,
    KeyEvent(Key, KeyState, KeyModifiers),
}

Events that the windowing system sends to Servo.

Variants

Idle

Sent when no message has arrived, but the event loop was kicked for some reason (perhaps by another Servo subsystem).

FIXME(pcwalton): This is kind of ugly and may not work well with multiprocess Servo. It's possible that this should be something like CompositorMessageWindowEvent(compositor_task::Msg) instead.

Refresh

Sent when part of the window is marked dirty and needs to be redrawn. Before sending this message, the window must make the same GL context as in PrepareRenderingEvent current.

InitializeCompositing

Sent to initialize the GL context. The windowing system must have a valid, current GL context when this message is sent.

Resize

Sent when the window is resized.

LoadUrl

Sent when a new URL is to be loaded.

MouseWindowEventClass

Sent when a mouse hit test is to be performed.

MouseWindowMoveEventClass

Sent when a mouse move.

Scroll

Sent when the user scrolls. The first point is the delta and the second point is the origin.

Zoom

Sent when the user zooms.

PinchZoom

Simulated "pinch zoom" gesture for non-touch platforms (e.g. ctrl-scrollwheel).

Navigation

Sent when the user uses chrome navigation (i.e. backspace or shift-backspace).

Quit

Sent when the user quits the application

KeyEvent

Sent when a key input state changes