Trait compositing::windowing::EmbedderMethods

pub trait EmbedderMethods {
    // Required method
    fn create_event_loop_waker(&mut self) -> Box<dyn EventLoopWaker>;

    // Provided methods
    fn register_webxr(&mut self, _: &mut MainThreadRegistry, _: EmbedderProxy) { ... }
    fn get_user_agent_string(&self) -> Option<String> { ... }
    fn get_version_string(&self) -> Option<String> { ... }
    fn get_protocol_handlers(&self) -> ProtocolRegistry { ... }
}

Required Methods

fn create_event_loop_waker(&mut self) -> Box<dyn EventLoopWaker>

Returns a thread-safe object to wake up the window’s event loop.

Provided Methods

fn register_webxr(&mut self, _: &mut MainThreadRegistry, _: EmbedderProxy)

Register services with a WebXR Registry.

fn get_user_agent_string(&self) -> Option<String>

Returns the user agent string to report in network requests.

fn get_version_string(&self) -> Option<String>

Returns the version string of this embedder.

fn get_protocol_handlers(&self) -> ProtocolRegistry

Returns the protocol handlers implemented by that embedder. They will be merged with the default internal ones.