Trait compositing::windowing::WindowMethods

pub trait WindowMethods {
    fn get_coordinates(&self) -> EmbedderCoordinates;

    fn set_animation_state(&self, _state: AnimationState);

    fn get_gl_context(&self) -> GlContext;

    fn get_native_display(&self) -> NativeDisplay;

    fn get_gl_api(&self) -> GlApi;

    fn webrender_surfman(&self) -> WebrenderSurfman;
}

Required methods

fn get_coordinates(&self) -> EmbedderCoordinates

Get the coordinates of the native window, the screen and the framebuffer.

fn set_animation_state(&self, _state: AnimationState)

Set whether the application is currently animating. Typically, when animations are active, the window will want to avoid blocking on UI events, and just run the event loop at the vsync interval.

fn get_gl_context(&self) -> GlContext

Get the media GL context

fn get_native_display(&self) -> NativeDisplay

Get the media native display

fn get_gl_api(&self) -> GlApi

Get the GL api

fn webrender_surfman(&self) -> WebrenderSurfman

Get the webrender surfman instance

Loading content...