Module compositing::compositor

Structs

HitTestResult

IOCompositor
NB: Never block on the constellation, because sometimes the constellation blocks on us.

PipelineDetails

RenderNotifier

RenderTargetInfo

ScrollZoomEvent

Enums

CompositeTarget

CompositingReason
Why we performed a composite. This is used for debugging.

CompositionRequest

NotReadyToPaint

ReadyState
Holds the state when running reftests that determines when it is safe to save the output image.

ShutdownState

UnableToComposite

Constants

BUFFER_MAP_SIZE

MAX_ZOOM

MIN_ZOOM

Traits

ConvertFragmentTypeFromWebRender

ConvertPipelineIdFromWebRender

ConvertStackingContextFromWebRender

Functions

find_layer_with_pipeline_and_layer_id_for_layer

initialize_png

reporter_name