Module compositing::compositor

Structs

FrameTreeId

IOCompositor

NB: Never block on the constellation, because sometimes the constellation blocks on us.

PipelineDetails

RootPipeline

ScrollZoomEvent

Enums

CompositeTarget

CompositingReason

Why we performed a composite. This is used for debugging.

CompositionRequest

LayerPixel

One pixel in layer coordinate space.

NotReadyToPaint

ReadyState

Holds the state when running reftests that determines when it is safe to save the output image.

ShutdownState

UnableToComposite

Constants

MAX_ZOOM

MIN_ZOOM

Traits

ConvertPipelineIdFromWebRender