Struct compositing::compositor::IOCompositor

pub struct IOCompositor<Window: WindowMethods + ?Sized> {

    pub window: Rc<Window>,
    port: CompositorReceiver,
    root_pipeline: Option<CompositionPipeline>,
    pipeline_details: HashMap<PipelineId, PipelineDetails>,
    scale: Scale<f32, LayerPixel, DevicePixel>,
    viewport_zoom: PinchZoomFactor,
    min_viewport_zoom: Option<PinchZoomFactor>,
    max_viewport_zoom: Option<PinchZoomFactor>,
    page_zoom: Scale<f32, CSSPixel, DeviceIndependentPixel>,
    composite_target: CompositeTarget,
    composition_request: CompositionRequest,
    pub shutdown_state: ShutdownState,
    zoom_action: bool,
    zoom_time: f64,
    frame_tree_id: FrameTreeId,
    constellation_chan: Sender<ConstellationMsg>,
    time_profiler_chan: ProfilerChan,
    touch_handler: TouchHandler,
    pending_scroll_zoom_events: Vec<ScrollZoomEvent>,
    waiting_for_results_of_scroll: bool,
    ready_to_save_state: ReadyState,
    webrender: Renderer,
    webrender_document: DocumentId,
    webrender_api: RenderApi,
    webrender_surfman: WebrenderSurfman,
    webrender_gl: Rc<dyn Gl>,
    pub webxr_main_thread: MainThreadRegistry,
    pending_paint_metrics: HashMap<PipelineId, Epoch>,
    embedder_coordinates: EmbedderCoordinates,
    cursor: Cursor,
    cursor_pos: DevicePoint,
    output_file: Option<String>,
    is_running_problem_test: bool,
    exit_after_load: bool,
    convert_mouse_to_touch: bool,
    waiting_on_pending_frame: bool,

}

NB: Never block on the constellation, because sometimes the constellation blocks on us.

Fields
window: Rc<Window>

The application window.

port: CompositorReceiver

The port on which we receive messages.

root_pipeline: Option<CompositionPipeline>

The root pipeline.

pipeline_details: HashMap<PipelineId, PipelineDetails>

Tracks details about each active pipeline that the compositor knows about.

scale: Scale<f32, LayerPixel, DevicePixel>

The scene scale, to allow for zooming and high-resolution painting.

viewport_zoom: PinchZoomFactor

“Mobile-style” zoom that does not reflow the page.

min_viewport_zoom: Option<PinchZoomFactor>

Viewport zoom constraints provided by @viewport.

max_viewport_zoom: Option<PinchZoomFactor>page_zoom: Scale<f32, CSSPixel, DeviceIndependentPixel>

“Desktop-style” zoom that resizes the viewport to fit the window.

composite_target: CompositeTarget

The type of composition to perform

composition_request: CompositionRequest

Tracks whether we should composite this frame.

shutdown_state: ShutdownState

Tracks whether we are in the process of shutting down, or have shut down and should close the compositor.

zoom_action: bool

Tracks whether the zoom action has happened recently.

zoom_time: f64

The time of the last zoom action has started.

frame_tree_id: FrameTreeId

The current frame tree ID (used to reject old paint buffers)

constellation_chan: Sender<ConstellationMsg>

The channel on which messages can be sent to the constellation.

time_profiler_chan: ProfilerChan

The channel on which messages can be sent to the time profiler.

touch_handler: TouchHandler

Touch input state machine

pending_scroll_zoom_events: Vec<ScrollZoomEvent>

Pending scroll/zoom events.

waiting_for_results_of_scroll: bool

Whether we’re waiting on a recomposite after dispatching a scroll.

ready_to_save_state: ReadyState

Used by the logic that determines when it is safe to output an image for the reftest framework.

webrender: Renderer

The webrender renderer.

webrender_document: DocumentId

The active webrender document.

webrender_api: RenderApi

The webrender interface, if enabled.

webrender_surfman: WebrenderSurfman

The surfman instance that webrender targets

webrender_gl: Rc<dyn Gl>

The GL bindings for webrender

webxr_main_thread: MainThreadRegistry

Some XR devices want to run on the main thread.

pending_paint_metrics: HashMap<PipelineId, Epoch>

Map of the pending paint metrics per layout thread. The layout thread for each specific pipeline expects the compositor to paint frames with specific given IDs (epoch). Once the compositor paints these frames, it records the paint time for each of them and sends the metric to the corresponding layout thread.

embedder_coordinates: EmbedderCoordinates

The coordinates of the native window, its view and the screen.

cursor: Cursor

Current mouse cursor.

cursor_pos: DevicePoint

Current cursor position.

output_file: Option<String>is_running_problem_test: boolexit_after_load: bool

True to exit after page load (‘-x’).

convert_mouse_to_touch: bool

True to translate mouse input into touch events.

waiting_on_pending_frame: bool

True if a WR frame render has been requested. Screenshots taken before the render is complete will not reflect the most up to date rendering.

Implementations

impl<Window: WindowMethods + ?Sized> IOCompositor<Window>

fn new(
window: Rc<Window>,
state: InitialCompositorState,
output_file: Option<String>,
is_running_problem_test: bool,
exit_after_load: bool,
convert_mouse_to_touch: bool
) -> Self

pub fn create(
window: Rc<Window>,
state: InitialCompositorState,
output_file: Option<String>,
is_running_problem_test: bool,
exit_after_load: bool,
convert_mouse_to_touch: bool
) -> Self

pub fn deinit(self)

pub fn update_cursor(&mut self, hit_test_results: HitTestResult)

pub fn maybe_start_shutting_down(&mut self)

fn start_shutting_down(&mut self)

fn finish_shutting_down(&mut self)

fn handle_browser_message(&mut self, msg: Msg) -> bool

fn handle_webrender_message(&mut self, msg: WebrenderMsg)

Accept messages from content processes that need to be relayed to the WebRender instance in the parent process.

fn change_running_animations_state(
&mut self,
pipeline_id: PipelineId,
animation_state: AnimationState
)

Sets or unsets the animations-running flag for the given pipeline, and schedules a recomposite if necessary.

fn pipeline_details(&mut self, pipeline_id: PipelineId) -> &mut PipelineDetails

pub fn pipeline(&self, pipeline_id: PipelineId) -> Option<&CompositionPipeline>

fn set_frame_tree(&mut self, frame_tree: &SendableFrameTree)

fn create_pipeline_details_for_frame_tree(
&mut self,
frame_tree: &SendableFrameTree
)

fn remove_pipeline_root_layer(&mut self, pipeline_id: PipelineId)

fn send_window_size(&self, size_type: WindowSizeType)

pub fn on_resize_window_event(&mut self) -> bool

pub fn on_mouse_window_event_class(
&mut self,
mouse_window_event: MouseWindowEvent
)

fn dispatch_mouse_window_event_class(
&mut self,
mouse_window_event: MouseWindowEvent
)

fn hit_test_at_point(&self, point: DevicePoint) -> HitTestResult

pub fn on_mouse_window_move_event_class(&mut self, cursor: DevicePoint)

fn dispatch_mouse_window_move_event_class(&mut self, cursor: DevicePoint)

fn send_touch_event(
&self,
event_type: TouchEventType,
identifier: TouchId,
point: DevicePoint
)

pub fn send_wheel_event(&mut self, delta: WheelDelta, point: DevicePoint)

pub fn on_touch_event(
&mut self,
event_type: TouchEventType,
identifier: TouchId,
location: DevicePoint
)

fn on_touch_down(&mut self, identifier: TouchId, point: DevicePoint)

fn on_touch_move(&mut self, identifier: TouchId, point: DevicePoint)

fn on_touch_up(&mut self, identifier: TouchId, point: DevicePoint)

fn on_touch_cancel(&mut self, identifier: TouchId, point: DevicePoint)

fn simulate_mouse_click(&mut self, p: DevicePoint)

http://w3c.github.io/touch-events/#mouse-events

pub fn on_wheel_event(&mut self, delta: WheelDelta, p: DevicePoint)

pub fn on_scroll_event(
&mut self,
delta: ScrollLocation,
cursor: DeviceIntPoint,
phase: TouchEventType
)

fn on_scroll_window_event(
&mut self,
scroll_location: ScrollLocation,
cursor: DeviceIntPoint
)

fn process_pending_scroll_events(&mut self)

fn process_animations(&mut self)

If there are any animations running, dispatches appropriate messages to the constellation.

fn tick_animations_for_pipeline(&mut self, pipeline_id: PipelineId)

fn constrain_viewport(
&mut self,
pipeline_id: PipelineId,
constraints: ViewportConstraints
)

fn hidpi_factor(&self) -> Scale<f32, DeviceIndependentPixel, DevicePixel>

fn device_pixels_per_page_px(&self) -> Scale<f32, CSSPixel, DevicePixel>

fn update_zoom_transform(&mut self)

pub fn on_zoom_reset_window_event(&mut self)

pub fn on_zoom_window_event(&mut self, magnification: f32)

fn update_page_zoom_for_webrender(&mut self)

pub fn on_pinch_zoom_window_event(&mut self, magnification: f32)

Simulate a pinch zoom

fn send_viewport_rects(&self)

fn animations_active(&self) -> bool

fn is_ready_to_paint_image_output(&mut self) -> Result<(), NotReadyToPaint>

Query the constellation to see if the current compositor output matches the current frame tree output, and if the associated script threads are idle.

pub fn composite(&mut self)

fn composite_specific_target(
&mut self,
target: CompositeTarget,
rect: Option<Rect<f32, CSSPixel>>
) -> Result<Option<Image>, UnableToComposite>

Composite either to the screen or to a png image or both. Returns Ok if composition was performed or Err if it was not possible to composite for some reason. If CompositeTarget is Window or Png no image data is returned; in the latter case the image is written directly to a file. If CompositeTarget is WindowAndPng Ok(Some(png::Image)) is returned.

fn composite_if_necessary(&mut self, reason: CompositingReason)

fn clear_background(&self)

fn assert_no_gl_error(&self)

fn assert_gl_framebuffer_complete(&self)

fn get_root_pipeline_id(&self) -> Option<PipelineId>

pub fn receive_messages(&mut self) -> bool

pub fn perform_updates(&mut self) -> bool

pub fn repaint_synchronously(&mut self)

Repaints and recomposites synchronously. You must be careful when calling this, as if a paint is not scheduled the compositor will hang forever.

This is used when resizing the window.

pub fn pinch_zoom_level(&self) -> f32

fn set_pinch_zoom_level(&mut self, zoom: f32)

pub fn toggle_webrender_debug(&mut self, option: WebRenderDebugOption)

pub fn capture_webrender(&mut self)