Enum compositing::compositor_thread::Msg

pub enum Msg {
    Exit,
    ShutdownComplete,
    GetNativeDisplay(Sender<Option<NativeDisplay>>),
    InitializeLayersForPipeline(PipelineId, Epoch, Vec<LayerProperties>),
    ScrollFragmentPoint(PipelineId, LayerId, Point2D<f32>, bool),
    AssignPaintedBuffers(PipelineId, Epoch, Vec<(LayerId, Box<LayerBufferSet>)>, FrameTreeId),
    ChangePageTitle(PipelineId, Option<String>),
    ChangePageUrl(PipelineId, Url),
    ChangeRunningAnimationsState(PipelineId, AnimationState),
    SetFrameTree(SendableFrameTree, IpcSender<()>),
    LoadStart(bool, bool),
    LoadComplete(bool, bool, bool),
    DelayedCompositionTimeout(u64),
    Recomposite(CompositingReason),
    KeyEvent(Key, KeyState, KeyModifiers),
    TouchEventProcessed(EventResult),
    SetCursor(Cursor),
    CreatePng(IpcSender<Option<Image>>),
    PaintThreadExited(PipelineId),
    ViewportConstrained(PipelineId, ViewportConstraints),
    IsReadyToSaveImageReply(bool),
    NewFavicon(Url),
    HeadParsed,
    ReturnUnusedNativeSurfaces(Vec<NativeSurface>),
    CollectMemoryReports(ReportsChan),
    Status(Option<String>),
    GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>),
    MoveTo(Point2D<i32>),
    ResizeTo(Size2D<u32>),
    GetScrollOffset(PipelineId, LayerId, IpcSender<Point2D<f32>>),
    PipelineVisibilityChanged(PipelineId, bool),
    PipelineExited(PipelineId, IpcSender<()>),
}

Messages from the painting thread and the constellation thread to the compositor thread.

Variants
Exit

Requests that the compositor shut down.

ShutdownComplete

Informs the compositor that the constellation has completed shutdown. Required because the constellation can have pending calls to make (e.g. SetFrameTree) at the time that we send it an ExitMsg.

GetNativeDisplay(Sender<Option<NativeDisplay>>)

Requests the compositor's graphics metadata. Graphics metadata is what the painter needs to create surfaces that the compositor can see. On Linux this is the X display; on Mac this is the pixel format.

InitializeLayersForPipeline(PipelineId, Epoch, Vec<LayerProperties>)

Tells the compositor to create or update the layers for a pipeline if necessary (i.e. if no layer with that ID exists).

ScrollFragmentPoint(PipelineId, LayerId, Point2D<f32>, bool)

Scroll a page in a window

AssignPaintedBuffers(PipelineId, Epoch, Vec<(LayerId, Box<LayerBufferSet>)>, FrameTreeId)

Requests that the compositor assign the painted buffers to the given layers.

ChangePageTitle(PipelineId, Option<String>)

Alerts the compositor that the current page has changed its title.

ChangePageUrl(PipelineId, Url)

Alerts the compositor that the current page has changed its URL.

ChangeRunningAnimationsState(PipelineId, AnimationState)

Alerts the compositor that the given pipeline has changed whether it is running animations.

SetFrameTree(SendableFrameTree, IpcSender<()>)

Replaces the current frame tree, typically called during main frame navigation.

LoadStart(bool, bool)

The load of a page has begun: (can go back, can go forward).

LoadComplete(bool, bool, bool)

The load of a page has completed: (can go back, can go forward, is root frame).

DelayedCompositionTimeout(u64)

We hit the delayed composition timeout. (See delayed_composition.rs.)

Recomposite(CompositingReason)

Composite.

KeyEvent(Key, KeyState, KeyModifiers)

Sends an unconsumed key event back to the compositor.

TouchEventProcessed(EventResult)

Script has handled a touch event, and either prevented or allowed default actions.

SetCursor(Cursor)

Changes the cursor.

CreatePng(IpcSender<Option<Image>>)

Composite to a PNG file and return the Image over a passed channel.

PaintThreadExited(PipelineId)

Informs the compositor that the paint thread for the given pipeline has exited.

ViewportConstrained(PipelineId, ViewportConstraints)

Alerts the compositor that the viewport has been constrained in some manner

IsReadyToSaveImageReply(bool)

A reply to the compositor asking if the output image is stable.

NewFavicon(Url)

A favicon was detected

HeadParsed

ReturnUnusedNativeSurfaces(Vec<NativeSurface>)

Signal that the paint thread ignored the paint requests that carried these native surfaces, so that they can be re-added to the surface cache.

CollectMemoryReports(ReportsChan)

Collect memory reports and send them back to the given mem::ReportsChan.

Status(Option<String>)

A status message to be displayed by the browser chrome.

GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>)

Get Window Informations size and position

MoveTo(Point2D<i32>)

Move the window to a point

ResizeTo(Size2D<u32>)

Resize the window to size

GetScrollOffset(PipelineId, LayerId, IpcSender<Point2D<f32>>)

Get scroll offset of a layer

PipelineVisibilityChanged(PipelineId, bool)

Pipeline visibility changed

PipelineExited(PipelineId, IpcSender<()>)

A pipeline was shut down.