Enum compositing::ConstellationMsg

pub enum ConstellationMsg {

    Exit,
    GetBrowsingContext(PipelineId, IpcSender<Option<BrowsingContextId>>),
    GetPipeline(BrowsingContextId, IpcSender<Option<PipelineId>>),
    GetFocusTopLevelBrowsingContext(IpcSender<Option<TopLevelBrowsingContextId>>),
    IsReadyToSaveImage(HashMap<PipelineId, Epoch>),
    Keyboard(KeyboardEvent),
    AllowNavigationResponse(PipelineId, bool),
    LoadUrl(TopLevelBrowsingContextId, ServoUrl),
    ClearCache,
    TraverseHistory(TopLevelBrowsingContextId, TraversalDirection),
    WindowSize(Option<TopLevelBrowsingContextId>, WindowSizeData, WindowSizeType),
    TickAnimation(PipelineId, AnimationTickType),
    WebDriverCommand(WebDriverCommandMsg),
    Reload(TopLevelBrowsingContextId),
    LogEntry(Option<TopLevelBrowsingContextId>, Option<String>, LogEntry),
    NewBrowser(ServoUrl, TopLevelBrowsingContextId),
    CloseBrowser(TopLevelBrowsingContextId),
    SendError(Option<TopLevelBrowsingContextId>, String),
    SelectBrowser(TopLevelBrowsingContextId),
    ForwardEvent(PipelineId, CompositorEvent),
    SetCursor(Cursor),
    EnableProfiler(Duration, Duration),
    DisableProfiler,
    ExitFullScreen(TopLevelBrowsingContextId),
    MediaSessionAction(MediaSessionActionType),
    ChangeBrowserVisibility(TopLevelBrowsingContextId, bool),
    IMEDismissed,

}

Messages to the constellation.

Variants

Exit

Exit the constellation.

GetBrowsingContext

Request that the constellation send the BrowsingContextId corresponding to the document with the provided pipeline id

Tuple Fields of GetBrowsingContext

0: PipelineId1: IpcSender<Option<BrowsingContextId>>

GetPipeline

Request that the constellation send the current pipeline id for the provided browsing context id, over a provided channel.

Tuple Fields of GetPipeline

0: BrowsingContextId1: IpcSender<Option<PipelineId>>

GetFocusTopLevelBrowsingContext

Request that the constellation send the current focused top-level browsing context id, over a provided channel.

Tuple Fields of GetFocusTopLevelBrowsingContext

0: IpcSender<Option<TopLevelBrowsingContextId>>

IsReadyToSaveImage

Query the constellation to see if the current compositor output is stable

Tuple Fields of IsReadyToSaveImage

0: HashMap<PipelineId, Epoch>

Keyboard

Inform the constellation of a key event.

Tuple Fields of Keyboard

0: KeyboardEvent

AllowNavigationResponse

Whether to allow script to navigate.

Tuple Fields of AllowNavigationResponse

0: PipelineId1: bool

LoadUrl

Request to load a page.

Tuple Fields of LoadUrl

0: TopLevelBrowsingContextId1: ServoUrl

ClearCache

Clear the network cache.

TraverseHistory

Request to traverse the joint session history of the provided browsing context.

Tuple Fields of TraverseHistory

0: TopLevelBrowsingContextId1: TraversalDirection

WindowSize

Inform the constellation of a window being resized.

Tuple Fields of WindowSize

0: Option<TopLevelBrowsingContextId>1: WindowSizeData2: WindowSizeType

TickAnimation

Requests that the constellation instruct layout to begin a new tick of the animation.

Tuple Fields of TickAnimation

0: PipelineId1: AnimationTickType

WebDriverCommand

Dispatch a webdriver command

Tuple Fields of WebDriverCommand

0: WebDriverCommandMsg

Reload

Reload a top-level browsing context.

Tuple Fields of Reload

0: TopLevelBrowsingContextId

LogEntry

A log entry, with the top-level browsing context id and thread name

Tuple Fields of LogEntry

0: Option<TopLevelBrowsingContextId>1: Option<String>2: LogEntry

NewBrowser

Create a new top level browsing context.

Tuple Fields of NewBrowser

0: ServoUrl1: TopLevelBrowsingContextId

CloseBrowser

Close a top level browsing context.

Tuple Fields of CloseBrowser

0: TopLevelBrowsingContextId

SendError

Panic a top level browsing context.

Tuple Fields of SendError

0: Option<TopLevelBrowsingContextId>1: String

SelectBrowser

Make browser visible.

Tuple Fields of SelectBrowser

0: TopLevelBrowsingContextId

ForwardEvent

Forward an event to the script task of the given pipeline.

Tuple Fields of ForwardEvent

0: PipelineId1: CompositorEvent

SetCursor

Requesting a change to the onscreen cursor.

Tuple Fields of SetCursor

0: Cursor

EnableProfiler

Enable the sampling profiler, with a given sampling rate and max total sampling duration.

Tuple Fields of EnableProfiler

0: Duration1: Duration

DisableProfiler

Disable the sampling profiler.

ExitFullScreen

Request to exit from fullscreen mode

Tuple Fields of ExitFullScreen

0: TopLevelBrowsingContextId

MediaSessionAction

Media session action.

Tuple Fields of MediaSessionAction

0: MediaSessionActionType

ChangeBrowserVisibility

Toggle browser visibility.

Tuple Fields of ChangeBrowserVisibility

0: TopLevelBrowsingContextId1: bool

IMEDismissed

Virtual keyboard was dismissed