Struct compositing::pipeline::Pipeline

pub struct Pipeline {
    pub id: PipelineId,
    pub parent_info: Option<(PipelineId, SubpageId)>,
    pub script_chan: IpcSender<ConstellationControlMsg>,
    pub layout_chan: LayoutControlChan,
    pub compositor_proxy: Box<CompositorProxy + 'static + Send>,
    pub chrome_to_paint_chan: Sender<ChromeToPaintMsg>,
    pub layout_shutdown_port: IpcReceiver<()>,
    pub paint_shutdown_port: IpcReceiver<()>,
    pub url: Url,
    pub title: Option<String>,
    pub size: Option<TypedSize2D<PagePx, f32>>,
    pub running_animations: bool,
    pub children: Vec<FrameId>,
    pub is_private: bool,
}

A uniquely-identifiable pipeline of script thread, layout thread, and paint thread.

Fields

id

parent_info

script_chan

layout_chan

A channel to layout, for performing reflows and shutdown.

compositor_proxy

A channel to the compositor.

chrome_to_paint_chan

layout_shutdown_port

paint_shutdown_port

url

URL corresponding to the most recently-loaded page.

title

The title of the most recently-loaded page.

size

running_animations

Whether this pipeline is currently running animations. Pipelines that are running animations cause composites to be continually scheduled.

children

is_private

Methods

impl Pipeline

fn create<LTF, STF>(state: InitialPipelineState) -> (Pipeline, UnprivilegedPipelineContent, PrivilegedPipelineContent) where LTF: LayoutThreadFactory, STF: ScriptThreadFactory

Starts a paint thread, layout thread, and possibly a script thread. Returns the channels wrapped in a struct.

fn new(id: PipelineId, parent_info: Option<(PipelineId, SubpageId)>, script_chan: IpcSender<ConstellationControlMsg>, layout_chan: LayoutControlChan, compositor_proxy: Box<CompositorProxy + 'static + Send>, chrome_to_paint_chan: Sender<ChromeToPaintMsg>, layout_shutdown_port: IpcReceiver<()>, paint_shutdown_port: IpcReceiver<()>, url: Url, size: Option<TypedSize2D<PagePx, f32>>) -> Pipeline

fn grant_paint_permission(&self)

fn revoke_paint_permission(&self)

fn exit(&self)

fn freeze(&self)

fn thaw(&self)

fn force_exit(&self)

fn to_sendable(&self) -> CompositionPipeline

fn add_child(&mut self, frame_id: FrameId)

fn remove_child(&mut self, frame_id: FrameId)

fn trigger_mozbrowser_event(&self, subpage_id: SubpageId, event: MozBrowserEvent)