Struct compositing::IOCompositor

pub struct IOCompositor<Window: WindowMethods> {
    window: Rc<Window>,
    native_display: Option<NativeDisplay>,
    port: Box<CompositorReceiver>,
    context: Option<RenderContext>,
    root_pipeline: Option<CompositionPipeline>,
    pipeline_details: HashMap<PipelineId, PipelineDetails>,
    scene: Scene<CompositorData>,
    window_size: TypedSize2D<DevicePixel, u32>,
    viewport: Option<(TypedPoint2D<DevicePixel, u32>, TypedSize2D<DevicePixel, u32>)>,
    viewport_zoom: ScaleFactor<PagePx, ViewportPx, f32>,
    min_viewport_zoom: Option<ScaleFactor<PagePx, ViewportPx, f32>>,
    max_viewport_zoom: Option<ScaleFactor<PagePx, ViewportPx, f32>>,
    page_zoom: ScaleFactor<ViewportPx, ScreenPx, f32>,
    scale_factor: ScaleFactor<ScreenPx, DevicePixel, f32>,
    channel_to_self: Box<CompositorProxy + Send>,
    delayed_composition_timer: DelayedCompositionTimerProxy,
    composite_target: CompositeTarget,
    composition_request: CompositionRequest,
    shutdown_state: ShutdownState,
    last_composite_time: u64,
    zoom_action: bool,
    zoom_time: f64,
    got_load_complete_message: bool,
    frame_tree_id: FrameTreeId,
    constellation_chan: Sender<ConstellationMsg>,
    time_profiler_chan: ProfilerChan,
    mem_profiler_chan: ProfilerChan,
    fragment_point: Option<Point2D<f32>>,
    touch_handler: TouchHandler,
    pending_scroll_zoom_events: Vec<ScrollZoomEvent>,
    ready_to_save_state: ReadyState,
    surface_map: SurfaceMap,
    pending_subpages: HashSet<PipelineId>,
    last_mouse_move_recipient: Option<PipelineId>,
    scroll_in_progress: bool,
    webrender: Option<Renderer>,
    webrender_api: Option<RenderApi>,
}

NB: Never block on the constellation, because sometimes the constellation blocks on us.

Fields
window: Rc<Window>

The application window.

native_display: Option<NativeDisplay>

The display this compositor targets. Will be None when using webrender.

port: Box<CompositorReceiver>

The port on which we receive messages.

context: Option<RenderContext>

The render context. This will be None if the windowing system has not yet sent us a PrepareRenderingEvent.

root_pipeline: Option<CompositionPipeline>

The root pipeline.

pipeline_details: HashMap<PipelineId, PipelineDetails>

Tracks details about each active pipeline that the compositor knows about.

scene: Scene<CompositorData>

The canvas to paint a page.

window_size: TypedSize2D<DevicePixel, u32>

The application window size.

viewport: Option<(TypedPoint2D<DevicePixel, u32>, TypedSize2D<DevicePixel, u32>)>

The overridden viewport.

viewport_zoom: ScaleFactor<PagePx, ViewportPx, f32>

"Mobile-style" zoom that does not reflow the page.

min_viewport_zoom: Option<ScaleFactor<PagePx, ViewportPx, f32>>

Viewport zoom constraints provided by @viewport.

max_viewport_zoom: Option<ScaleFactor<PagePx, ViewportPx, f32>> page_zoom: ScaleFactor<ViewportPx, ScreenPx, f32>

"Desktop-style" zoom that resizes the viewport to fit the window. See ViewportPx docs in util/geom.rs for details.

scale_factor: ScaleFactor<ScreenPx, DevicePixel, f32>

The device pixel ratio for this window.

channel_to_self: Box<CompositorProxy + Send> delayed_composition_timer: DelayedCompositionTimerProxy

A handle to the delayed composition timer.

composite_target: CompositeTarget

The type of composition to perform

composition_request: CompositionRequest

Tracks whether we should composite this frame.

shutdown_state: ShutdownState

Tracks whether we are in the process of shutting down, or have shut down and should close the compositor.

last_composite_time: u64

Tracks the last composite time.

zoom_action: bool

Tracks whether the zoom action has happened recently.

zoom_time: f64

The time of the last zoom action has started.

got_load_complete_message: bool

Whether the page being rendered has loaded completely. Differs from ReadyState because we can finish loading (ready) many times for a single page.

frame_tree_id: FrameTreeId

The current frame tree ID (used to reject old paint buffers)

constellation_chan: Sender<ConstellationMsg>

The channel on which messages can be sent to the constellation.

time_profiler_chan: ProfilerChan

The channel on which messages can be sent to the time profiler.

mem_profiler_chan: ProfilerChan

The channel on which messages can be sent to the memory profiler.

fragment_point: Option<Point2D<f32>>

Pending scroll to fragment event, if any

touch_handler: TouchHandler

Touch input state machine

pending_scroll_zoom_events: Vec<ScrollZoomEvent>

Pending scroll/zoom events.

ready_to_save_state: ReadyState

Used by the logic that determines when it is safe to output an image for the reftest framework.

surface_map: SurfaceMap

A data structure to cache unused NativeSurfaces.

pending_subpages: HashSet<PipelineId>

Pipeline IDs of subpages that the compositor has seen in a layer tree but which have not yet been painted.

last_mouse_move_recipient: Option<PipelineId>

The id of the pipeline that was last sent a mouse move event, if any.

scroll_in_progress: bool

Whether a scroll is in progress; i.e. whether the user's fingers are down.

webrender: Option<Renderer>

The webrender renderer, if enabled.

webrender_api: Option<RenderApi>

The webrender interface, if enabled.

Methods

impl<Window: WindowMethods> IOCompositor<Window>

fn new(window: Rc<Window>, state: InitialCompositorState) -> IOCompositor<Window>

fn create(window: Rc<Window>, state: InitialCompositorState) -> IOCompositor<Window>

fn start_shutting_down(&mut self)

fn finish_shutting_down(&mut self)

fn handle_browser_message(&mut self, msg: Msg) -> bool

fn change_running_animations_state(&mut self, pipeline_id: PipelineId, animation_state: AnimationState)

Sets or unsets the animations-running flag for the given pipeline, and schedules a recomposite if necessary.

fn pipeline_details(&mut self, pipeline_id: PipelineId) -> &mut PipelineDetails

fn pipeline(&self, pipeline_id: PipelineId) -> Option<&CompositionPipeline>

fn change_page_title(&mut self, pipeline_id: PipelineId, title: Option<String>)

fn change_page_url(&mut self, _: PipelineId, url: Url)

fn set_frame_tree(&mut self, frame_tree: &SendableFrameTree, response_chan: IpcSender<()>)

fn create_root_layer_for_pipeline_and_size(&mut self, pipeline: &CompositionPipeline, frame_size: Option<TypedSize2D<PagePx, f32>>) -> Rc<Layer<CompositorData>>

fn create_pipeline_details_for_frame_tree(&mut self, frame_tree: &SendableFrameTree)

fn find_pipeline_root_layer(&self, pipeline_id: PipelineId) -> Option<Rc<Layer<CompositorData>>>

fn collect_old_layers(&mut self, pipeline_id: PipelineId, new_layers: &[LayerProperties])

fn remove_pipeline_root_layer(&mut self, pipeline_id: PipelineId)

fn update_layer_if_exists(&mut self, pipeline_id: PipelineId, properties: LayerProperties) -> bool

fn create_or_update_base_layer(&mut self, pipeline_id: PipelineId, layer_properties: LayerProperties)

fn create_or_update_descendant_layer(&mut self, pipeline_id: PipelineId, layer_properties: LayerProperties)

fn create_descendant_layer(&mut self, pipeline_id: PipelineId, layer_properties: LayerProperties)

fn create_root_layer_for_subpage_if_necessary(&mut self, layer_properties: LayerProperties, parent_layer: Rc<Layer<CompositorData>>)

fn send_window_size(&self, size_type: WindowSizeType)

fn update_subpage_size_if_necessary(&self, layer_properties: &LayerProperties)

Sends the size of the given subpage up to the constellation. This will often trigger a reflow of that subpage.

fn move_layer(&self, pipeline_id: PipelineId, layer_id: LayerId, origin: TypedPoint2D<LayerPixel, f32>) -> bool

fn scroll_layer_to_fragment_point_if_necessary(&mut self, pipeline_id: PipelineId, layer_id: LayerId)

fn schedule_delayed_composite_if_necessary(&mut self)

fn assign_painted_buffers(&mut self, pipeline_id: PipelineId, layer_id: LayerId, new_layer_buffer_set: Box<LayerBufferSet>, epoch: Epoch, frame_tree_id: FrameTreeId)

fn assign_painted_buffers_to_layer(&mut self, layer: Rc<Layer<CompositorData>>, new_layer_buffer_set: Box<LayerBufferSet>, epoch: Epoch)

fn scroll_fragment_to_point(&mut self, pipeline_id: PipelineId, layer_id: LayerId, point: Point2D<f32>)

fn handle_window_message(&mut self, event: WindowEvent)

fn on_resize_window_event(&mut self, new_size: TypedSize2D<DevicePixel, u32>)

fn on_load_url_window_event(&mut self, url_string: String)

fn on_mouse_window_event_class(&mut self, mouse_window_event: MouseWindowEvent)

fn on_mouse_window_move_event_class(&mut self, cursor: TypedPoint2D<DevicePixel, f32>)

fn on_touch_down(&mut self, identifier: TouchId, point: TypedPoint2D<DevicePixel, f32>)

fn on_touch_move(&mut self, identifier: TouchId, point: TypedPoint2D<DevicePixel, f32>)

fn on_touch_up(&mut self, identifier: TouchId, point: TypedPoint2D<DevicePixel, f32>)

fn on_touch_cancel(&mut self, identifier: TouchId, point: TypedPoint2D<DevicePixel, f32>)

fn simulate_mouse_click(&self, p: TypedPoint2D<DevicePixel, f32>)

http://w3c.github.io/touch-events/#mouse-events

fn on_scroll_window_event(&mut self, delta: TypedPoint2D<DevicePixel, f32>, cursor: TypedPoint2D<DevicePixel, i32>)

fn on_scroll_start_window_event(&mut self, delta: TypedPoint2D<DevicePixel, f32>, cursor: TypedPoint2D<DevicePixel, i32>)

fn on_scroll_end_window_event(&mut self, delta: TypedPoint2D<DevicePixel, f32>, cursor: TypedPoint2D<DevicePixel, i32>)

fn process_pending_scroll_events(&mut self)

fn send_updated_display_ports_to_layout(&mut self)

Computes new display ports for each layer, taking the scroll position into account, and sends them to layout as necessary. This ultimately triggers a rerender of the content.

fn perform_updates_after_scroll(&mut self)

Performs buffer requests and starts the scrolling timer or schedules a recomposite as necessary.

fn process_animations(&mut self)

If there are any animations running, dispatches appropriate messages to the constellation.

fn tick_animations_for_pipeline(&mut self, pipeline_id: PipelineId)

fn constrain_viewport(&mut self, pipeline_id: PipelineId, constraints: ViewportConstraints)

fn device_pixels_per_screen_px(&self) -> ScaleFactor<ScreenPx, DevicePixel, f32>

fn device_pixels_per_page_px(&self) -> ScaleFactor<PagePx, DevicePixel, f32>

fn update_zoom_transform(&mut self)

fn on_zoom_reset_window_event(&mut self)

fn on_zoom_window_event(&mut self, magnification: f32)

fn on_pinch_zoom_window_event(&mut self, magnification: f32)

Simulate a pinch zoom

fn on_navigation_window_event(&self, direction: WindowNavigateMsg)

fn on_touchpad_pressure_event(&self, cursor: TypedPoint2D<DevicePixel, f32>, pressure: f32, phase: TouchpadPressurePhase)

fn on_key_event(&self, key: Key, state: KeyState, modifiers: KeyModifiers)

fn fill_paint_request_with_cached_layer_buffers(&mut self, paint_request: &mut PaintRequest)

fn convert_buffer_requests_to_pipeline_requests_map(&mut self, requests: Vec<(Rc<Layer<CompositorData>>, Vec<BufferRequest>)>) -> HashMap<PipelineId, Vec<PaintRequest>>

fn send_viewport_rect_for_layer(&self, layer: Rc<Layer<CompositorData>>)

fn send_viewport_rects_for_all_layers(&self)

fn send_webrender_viewport_rects(&self)

fn send_buffer_requests_for_all_layers(&mut self) -> bool

Returns true if any buffer requests were sent or false otherwise.

fn does_layer_have_outstanding_paint_messages(&self, layer: &Rc<Layer<CompositorData>>) -> bool

Check if a layer (or its children) have any outstanding paint results to arrive yet.

fn animations_active(&self) -> bool

fn is_ready_to_paint_image_output(&mut self) -> Result<(), NotReadyToPaint>

Query the constellation to see if the current compositor output matches the current frame tree output, and if the associated script threads are idle.

fn composite(&mut self)

fn composite_specific_target(&mut self, target: CompositeTarget) -> Result<Option<Image>, UnableToComposite>

Composite either to the screen or to a png image or both. Returns Ok if composition was performed or Err if it was not possible to composite for some reason. If CompositeTarget is Window or Png no image data is returned; in the latter case the image is written directly to a file. If CompositeTarget is WindowAndPng Ok(Some(png::Image)) is returned.

fn draw_img(&self, render_target_info: RenderTargetInfo, width: usize, height: usize) -> RgbImage

fn composite_if_necessary(&mut self, reason: CompositingReason)

fn composite_if_necessary_if_not_using_webrender(&mut self, reason: CompositingReason)

fn initialize_compositing(&mut self)

fn find_topmost_layer_at_point_for_layer(&self, layer: Rc<Layer<CompositorData>>, point_in_parent_layer: TypedPoint2D<LayerPixel, f32>, clip_rect_in_parent_layer: &TypedRect<LayerPixel, f32>) -> Option<HitTestResult>

fn find_topmost_layer_at_point(&self, point: TypedPoint2D<LayerPixel, f32>) -> Option<HitTestResult>

fn find_layer_with_pipeline_and_layer_id(&self, pipeline_id: PipelineId, layer_id: LayerId) -> Option<Rc<Layer<CompositorData>>>

fn cache_unused_buffers<B>(&mut self, buffers: B) where B: IntoIterator<Item=Box<LayerBuffer>>

fn get_root_pipeline_id(&self) -> Option<PipelineId>

fn dump_layer_tree(&self)

fn dump_layer_tree_layer(&self, layer: &Layer<CompositorData>, print_tree: &mut PrintTree)

fn start_scrolling_bounce_if_necessary(&mut self)

fn handle_events(&mut self, messages: Vec<WindowEvent>) -> bool

fn repaint_synchronously(&mut self)

Repaints and recomposites synchronously. You must be careful when calling this, as if a paint is not scheduled the compositor will hang forever.

This is used when resizing the window.

fn pinch_zoom_level(&self) -> f32

fn title_for_main_frame(&self)