Enum compositing::windowing::WindowEvent

pub enum WindowEvent {
    Idle,
    Refresh,
    InitializeCompositing,
    Resize(TypedSize2D<DevicePixel, u32>),
    TouchpadPressure(TypedPoint2D<DevicePixel, f32>, f32, TouchpadPressurePhase),
    Viewport(TypedPoint2D<DevicePixel, u32>, TypedSize2D<DevicePixel, u32>),
    LoadUrl(String),
    MouseWindowEventClass(MouseWindowEvent),
    MouseWindowMoveEventClass(TypedPoint2D<DevicePixel, f32>),
    Touch(TouchEventType, TouchId, TypedPoint2D<DevicePixel, f32>),
    Scroll(TypedPoint2D<DevicePixel, f32>, TypedPoint2D<DevicePixel, i32>, TouchEventType),
    Zoom(f32),
    PinchZoom(f32),
    ResetZoom,
    Navigation(WindowNavigateMsg),
    Quit,
    KeyEvent(Key, KeyState, KeyModifiers),
    Reload,
}

Events that the windowing system sends to Servo.

Variants
Idle

Sent when no message has arrived, but the event loop was kicked for some reason (perhaps by another Servo subsystem).

FIXME(pcwalton): This is kind of ugly and may not work well with multiprocess Servo. It's possible that this should be something like CompositorMessageWindowEvent(compositor_thread::Msg) instead.

Refresh

Sent when part of the window is marked dirty and needs to be redrawn. Before sending this message, the window must make the same GL context as in PrepareRenderingEvent current.

InitializeCompositing

Sent to initialize the GL context. The windowing system must have a valid, current GL context when this message is sent.

Resize(TypedSize2D<DevicePixel, u32>)

Sent when the window is resized.

TouchpadPressure(TypedPoint2D<DevicePixel, f32>, f32, TouchpadPressurePhase)

Touchpad Pressure

Viewport(TypedPoint2D<DevicePixel, u32>, TypedSize2D<DevicePixel, u32>)

Sent when you want to override the viewport.

LoadUrl(String)

Sent when a new URL is to be loaded.

MouseWindowEventClass(MouseWindowEvent)

Sent when a mouse hit test is to be performed.

MouseWindowMoveEventClass(TypedPoint2D<DevicePixel, f32>)

Sent when a mouse move.

Touch(TouchEventType, TouchId, TypedPoint2D<DevicePixel, f32>)

Touch event: type, identifier, point

Scroll(TypedPoint2D<DevicePixel, f32>, TypedPoint2D<DevicePixel, i32>, TouchEventType)

Sent when the user scrolls. The first point is the delta and the second point is the origin.

Zoom(f32)

Sent when the user zooms.

PinchZoom(f32)

Simulated "pinch zoom" gesture for non-touch platforms (e.g. ctrl-scrollwheel).

ResetZoom

Sent when the user resets zoom to default.

Navigation(WindowNavigateMsg)

Sent when the user uses chrome navigation (i.e. backspace or shift-backspace).

Quit

Sent when the user quits the application

KeyEvent(Key, KeyState, KeyModifiers)

Sent when a key input state changes

Reload

Sent when Ctr+R/Apple+R is called to reload the current page.