Trait compositing::windowing::WindowMethods

pub trait WindowMethods {
    fn framebuffer_size(&self) -> TypedSize2D<DevicePixel, u32>;
    fn size(&self) -> TypedSize2D<ScreenPx, f32>;
    fn present(&self);
    fn client_window(&self) -> (Size2D<u32>, Point2D<i32>);
    fn set_inner_size(&self, size: Size2D<u32>);
    fn set_position(&self, point: Point2D<i32>);
    fn set_page_title(&self, title: Option<String>);
    fn set_page_url(&self, url: Url);
    fn status(&self, Option<String>);
    fn load_start(&self, back: bool, forward: bool);
    fn load_end(&self, back: bool, forward: bool, root: bool);
    fn load_error(&self, code: NetError, url: String);
    fn head_parsed(&self);
    fn hidpi_factor(&self) -> ScaleFactor<ScreenPx, DevicePixel, f32>;
    fn native_display(&self) -> NativeDisplay;
    fn create_compositor_channel(&self) -> (Box<CompositorProxy + Send>, Box<CompositorReceiver>);
    fn prepare_for_composite(&self, width: usize, height: usize) -> bool;
    fn set_cursor(&self, cursor: Cursor);
    fn handle_key(&self, key: Key, mods: KeyModifiers);
    fn supports_clipboard(&self) -> bool;
    fn set_favicon(&self, url: Url);
}

Required Methods

fn framebuffer_size(&self) -> TypedSize2D<DevicePixel, u32>

Returns the size of the window in hardware pixels.

fn size(&self) -> TypedSize2D<ScreenPx, f32>

Returns the size of the window in density-independent "px" units.

fn present(&self)

Presents the window to the screen (perhaps by page flipping).

fn client_window(&self) -> (Size2D<u32>, Point2D<i32>)

Return the size of the window with head and borders and position of the window values

fn set_inner_size(&self, size: Size2D<u32>)

Set the size inside of borders and head

fn set_position(&self, point: Point2D<i32>)

Set the window position

fn set_page_title(&self, title: Option<String>)

Sets the page title for the current page.

fn set_page_url(&self, url: Url)

Sets the load data for the current page.

fn status(&self, Option<String>)

Called when the browser chrome should display a status message.

fn load_start(&self, back: bool, forward: bool)

Called when the browser has started loading a frame.

fn load_end(&self, back: bool, forward: bool, root: bool)

Called when the browser is done loading a frame.

fn load_error(&self, code: NetError, url: String)

Called when the browser encounters an error while loading a URL

fn head_parsed(&self)

Called when the

fn hidpi_factor(&self) -> ScaleFactor<ScreenPx, DevicePixel, f32>

Returns the hidpi factor of the monitor.

fn native_display(&self) -> NativeDisplay

Gets the OS native graphics display for this window.

fn create_compositor_channel(&self) -> (Box<CompositorProxy + Send>, Box<CompositorReceiver>)

Creates a channel to the compositor. The dummy parameter is needed because we don't have UFCS in Rust yet.

This is part of the windowing system because its implementation often involves OS-specific magic to wake the up window's event loop.

fn prepare_for_composite(&self, width: usize, height: usize) -> bool

Requests that the window system prepare a composite. Typically this will involve making some type of platform-specific graphics context current. Returns true if the composite may proceed and false if it should not.

fn set_cursor(&self, cursor: Cursor)

Sets the cursor to be used in the window.

fn handle_key(&self, key: Key, mods: KeyModifiers)

Process a key event.

fn supports_clipboard(&self) -> bool

Does this window support a clipboard

fn set_favicon(&self, url: Url)

Add a favicon