Trait compositing::windowing::WindowMethods

pub trait WindowMethods {
    fn present(&self);

    fn make_gl_context_current(&self);

    fn gl(&self) -> Rc<dyn Gl>;

    fn get_coordinates(&self) -> EmbedderCoordinates;

    fn set_animation_state(&self, _state: AnimationState);

    fn get_gl_context(&self) -> GlContext;

    fn get_native_display(&self) -> NativeDisplay;

    fn get_gl_api(&self) -> GlApi;
}

Required methods

fn present(&self)

Presents the window to the screen (perhaps by page flipping).

fn make_gl_context_current(&self)

Make the OpenGL context current.

fn gl(&self) -> Rc<dyn Gl>

Return the GL function pointer trait.

fn get_coordinates(&self) -> EmbedderCoordinates

Get the coordinates of the native window, the screen and the framebuffer.

fn set_animation_state(&self, _state: AnimationState)

Set whether the application is currently animating. Typically, when animations are active, the window will want to avoid blocking on UI events, and just run the event loop at the vsync interval.

fn get_gl_context(&self) -> GlContext

Get the GL context

fn get_native_display(&self) -> NativeDisplay

Get the native display

fn get_gl_api(&self) -> GlApi

Get the GL api

Loading content...