Trait compositing::windowing::WindowMethods

pub trait WindowMethods {
    // Required methods
    fn get_coordinates(&self) -> EmbedderCoordinates;
    fn set_animation_state(&self, _state: AnimationState);
    fn rendering_context(&self) -> RenderingContext;
}

Required Methods

fn get_coordinates(&self) -> EmbedderCoordinates

Get the coordinates of the native window, the screen and the framebuffer.

fn set_animation_state(&self, _state: AnimationState)

Set whether the application is currently animating. Typically, when animations are active, the window will want to avoid blocking on UI events, and just run the event loop at the vsync interval.

fn rendering_context(&self) -> RenderingContext

Get the RenderingContext of this Window.