Trait compositing::windowing::WindowMethods

pub trait WindowMethods {
    fn framebuffer_size(&self) -> TypedSize2D<DevicePixel, u32>;
    fn size(&self) -> TypedSize2D<ScreenPx, f32>;
    fn present(&self);
    fn set_ready_state(&self, ready_state: ReadyState);
    fn set_paint_state(&self, paint_state: PaintState);
    fn set_page_title(&self, title: Option<String>);
    fn set_page_url(&self, url: Url);
    fn load_end(&self);
    fn hidpi_factor(&self) -> ScaleFactor<ScreenPx, DevicePixel, f32>;
    fn native_metadata(&self) -> NativeGraphicsMetadata;
    fn create_compositor_channel(_: &Option<Rc<Self>>) -> (Box<CompositorProxy + Send>, Box<CompositorReceiver>);
    fn prepare_for_composite(&self) -> bool;
    fn set_cursor(&self, cursor: Cursor);
    fn handle_key(&self, key: Key, mods: KeyModifiers);
}

Required Methods

fn framebuffer_size(&self) -> TypedSize2D<DevicePixel, u32>

Returns the size of the window in hardware pixels.

fn size(&self) -> TypedSize2D<ScreenPx, f32>

Returns the size of the window in density-independent "px" units.

fn present(&self)

Presents the window to the screen (perhaps by page flipping).

fn set_ready_state(&self, ready_state: ReadyState)

Sets the ready state of the current page.

fn set_paint_state(&self, paint_state: PaintState)

Sets the paint state of the current page.

fn set_page_title(&self, title: Option<String>)

Sets the page title for the current page.

fn set_page_url(&self, url: Url)

Sets the load data for the current page.

fn load_end(&self)

Called when the browser is done loading a frame.

fn hidpi_factor(&self) -> ScaleFactor<ScreenPx, DevicePixel, f32>

Returns the hidpi factor of the monitor.

fn native_metadata(&self) -> NativeGraphicsMetadata

Gets the OS native graphics information for this window.

fn create_compositor_channel(_: &Option<Rc<Self>>) -> (Box<CompositorProxy + Send>, Box<CompositorReceiver>)

Creates a channel to the compositor. The dummy parameter is needed because we don't have UFCS in Rust yet.

This is part of the windowing system because its implementation often involves OS-specific magic to wake the up window's event loop.

fn prepare_for_composite(&self) -> bool

Requests that the window system prepare a composite. Typically this will involve making some type of platform-specific graphics context current. Returns true if the composite may proceed and false if it should not.

fn set_cursor(&self, cursor: Cursor)

Sets the cursor to be used in the window.

fn handle_key(&self, key: Key, mods: KeyModifiers)

Process a key event.